requirements.txt
openai
twilio
feedparser
python-dotenv
